+++ crates/telio-nurse/src/data.rs
@@ -1,3 +1,4 @@
+use std::collections::{BTreeSet, HashMap};
 use telio_crypto::PublicKey;
 use telio_model::config::Config;
 
@@ -11,9 +12,9 @@ pub struct HeartbeatInfo {
     /// Connectivity matrix of the meshnet
     pub connectivity_matrix: String,
     /// Public keys of internal nodes
-    pub internal_sorted_public_keys: Vec<PublicKey>,
+    pub internal_sorted_public_keys: BTreeSet<PublicKey>,
     /// Public keys of external nodes
-    pub external_sorted_public_keys: Vec<PublicKey>,
+    pub external_sorted_public_keys: BTreeSet<PublicKey>,
     /// How often to send heartbeats
     pub heartbeat_interval: i32,
     /// String with comma-separated list of `meshnet_id:fingerprint:connection_state` for all external nodes
@@ -32,19 +33,18 @@ pub enum AnalyticsMessage {
 /// Represents an update to the meshnet config
 #[derive(Clone, Default)]
 pub struct MeshConfigUpdateEvent {
-    /// All local nodes from the config passed to telio library.
-    pub local_nodes: Vec<PublicKey>,
+    /// All the nodes (public_key, is_local) from the config passed to telio library.
+    pub nodes: HashMap<PublicKey, bool>,
 }
 
 impl From<&Config> for MeshConfigUpdateEvent {
     fn from(config: &Config) -> Self {
         if let Some(peers) = &config.peers {
-            let local_nodes = peers
+            let nodes = peers
                 .iter()
-                .filter(|p| p.is_local)
-                .map(|p| p.public_key)
-                .collect::<Vec<_>>();
-            MeshConfigUpdateEvent { local_nodes }
+                .map(|p| (p.public_key, p.is_local))
+                .collect::<HashMap<_, _>>();
+            MeshConfigUpdateEvent { nodes }
         } else {
             MeshConfigUpdateEvent::default()
         }